Allow php7.x to generate php5 compatible code for type double with default '0'
When generating classes in php7.x the output was different when compared
to the php5.5 generated code resulting in failure when running the code
on the old interpreter.

The fix might not be needed once all infrastructure is on 7.x, but for
compatible switches between php versions during the migration phase
it makes sense to keep the generated output in sync independent of the
php version used to generate.